Fix sortable ontologies list
